Changed model back to Float32

+++ utils.py
@@ -18,7 +18,7 @@ def make_dist_sampler(
 
     if distribution.lower() == "normal":
         dist_sampler = lambda x: (
-            torch.empty(x, device=device, dtype=torch.float64).normal_(mean=0, std=1)
+            torch.empty(x, device=device, dtype=torch.float32).normal_(mean=0, std=1)
         )
 
     elif distribution.lower() == "bernoulli":
@@ -123,7 +123,7 @@ def load_dataset(dataset_importer, device, fltype, validation, mean, std):
             train_dataset = torchvision.datasets.ImageFolder(train_datasetpath)
             test_dataset = torchvision.datasets.ImageFolder(test_datasetpath)
 
-            x_test = np.empty((len(test_dataset.targets), 3, 64, 64), dtype=np.float64)
+            x_test = np.empty((len(test_dataset.targets), 3, 64, 64), dtype=np.float32)
             y_test = np.empty((len(test_dataset.targets)))
             for indx, (img, label) in enumerate(test_dataset.imgs):
                 x_test[indx] = torchvision.transforms.ToTensor()(
@@ -138,7 +138,7 @@ def load_dataset(dataset_importer, device, fltype, validation, mean, std):
             np.save("./datasets/tiny-imagenet-200/y_test.npy", y_test)
 
             x_train = np.empty(
-                (len(train_dataset.targets), 3, 64, 64), dtype=np.float64
+                (len(train_dataset.targets), 3, 64, 64), dtype=np.float32
             )
             y_train = np.empty((len(train_dataset.targets)))
             for indx, (img, label) in enumerate(train_dataset.imgs):
@@ -321,7 +321,7 @@ def construct_dataloaders(
             x_train, y_train, x_test, y_test = load_dataset(
                 tv_dataset,
                 device,
-                torch.float64,
+                torch.float32,
                 validation=validation,
                 mean=mean,
                 std=std,
@@ -524,7 +524,7 @@ def train(
         onehots = (
             torch.nn.functional.one_hot(target, num_classes).to(device).to(data.dtype)
         )
-        data, target = data.to(device).to(torch.double), target.to(device)
+        data, target = data.to(device), target.to(device)
         if (
             batch_idx == len(train_loader) - 2 or batch_idx == len(train_loader) - 3
         ) and comp_angles:
@@ -591,7 +591,7 @@ def test(
                 .to(device)
                 .to(data.dtype)
             )
-            data, target = data.to(device).to(torch.double), target.to(device)
+            data, target = data.to(device), target.to(device)
 
             loss, output = model.test_step(data, target, onehots, loss_func)
             test_loss += loss
